+++ boden/include/bdn/ios/WindowCore.hh
@@ -108,7 +108,7 @@ public:
     
     
     
-    Size calcPreferredSize(int availableWidth=-1, int availableHeight=-1) const override
+    Size calcPreferredSize(double availableWidth=-1, double availableHeight=-1) const override
     {
         // the implementation for this must be provided by the outer Window object.
         throw NotImplementedError("WindowCore::calcPreferredSize");
