+++ boden/include/bdn/test/MockTextViewCore.h
@@ -48,7 +48,7 @@ public:
 
 
 	
-	Size calcPreferredSize(int availableWidth=-1, int availableHeight=-1) const override
+	Size calcPreferredSize(double availableWidth=-1, double availableHeight=-1) const override
 	{
 		BDN_REQUIRE_IN_MAIN_THREAD();
